userspace + fat32 added

+++ inc/drivers/idt.h
@@ -134,7 +134,7 @@ struct madt_local_x2apic {
 };
 
 typedef struct {
-    uint8_t arg[8];
+    uint64_t arg[8];
 } arg;
 void sleep_ms(uint64_t ms);
 void parse_madt(struct acpi_table_madt *madt);
